feat: complete client package

+++ pkg/client/response.go
@@ -2,5 +2,5 @@ package client
 
 type Response struct {
 	Body       []byte
-	StatusCode int64
+	StatusCode int
 }
